Set defaults for smagic

Se the default values for a few arguments of the function smagic().

setSmagic(
  .smagic.class = FALSE,
  .delim = c("{", "}"),
  .sep = "",
  .data.table = TRUE,
  reset = FALSE
)

Arguments

.smagic.class

Logical scalar, default is FALSE. If TRUE, the vector returned will be of class smagic. This class has a simple print method which is useful to display small vectors.

.delim

Character vector of length 1 or 2. Default is c("{", "}"). Defines the opening and the closing delimiters for interpolation.

If of length 1, it must be of the form: 1) the opening delimiter, 2) a single space, 3) the closing delimiter. Ex: ".[ ]" is equivalent to c(".[", "]"). The default value is equivalent to "{ }".

[ ]: R:%20 [", "]: R:%22,%20%22

.sep

Character scalar, default is the empty string "". It is used to collapse all the elements in ... before applying any operation.

.data.table

Logical, default is TRUE. If you use smagic in a data.table call and interpolate variables within the data.table, you want this argument to be TRUE. It only incurs a small overhead.

reset

Logical scalar, default is FALSE. Whether to reset all values.

Details

By default, each call to setSmagic adds modifications to the default values. To set a few default values and resetting the others, you need to use reset = TRUE.

Author

Laurent Berge

Examples

# we change the default display of the results of smagic
setSmagic(.smagic.class = TRUE)
smagic("{S!x, y}{2 each?1:2}")
#> x1
#> y1
#> x2
#> y2

# back to a regular character vector
setSmagic(reset = TRUE)
smagic("{S!x, y}{2 each?1:2}")
#> [1] "x1" "y1" "x2" "y2"