Chains basic operations to character vectors

Simple tool to perform multiple operations to character vectors.

string_ops_alias(op = NULL, pre_unik = NULL, namespace = NULL)

string_ops(x, op, pre_unik = NULL, namespace = NULL, envir = parent.frame())

Arguments

op

Character scalar. Character scalar containing the comma separated values of operations to perform to the vector. The 50+ operations are detailed in the help page of smagic().

pre_unik

Logical scalar, default is NULL. Whether to first unique the vector before applying the possibly costly string operations, and merging back the result. For very large vectors with repeated values the time gained can be substantial. By default, this is TRUE for vector of length 1M or more.

namespace

Character scalar or NULL (default). Only useful for package developers. As a regular end-user you shouldn't care! If your package uses smagic, you should care. It is useful only if your package uses 'custom' smagic operations, set with smagic_register_fun() or smagic_register_ops().

If so pass the name of your package in this argument so that your function can access the new smagic operations defined within your package.

x

A character vector. If not a character vector but atomistic (i.e. not a list), it will be converted to a character vector.

envir

Environment in which to evaluate the interpolations if the flag "magic" is provided. Default is parent.frame().

Value

In general it returns a character vector. It may be of a length different from the original one, depending on the operations performed.

Details

This function is a simple wrapper around smagic. Formally, string_ops(x, "op1, op2") is equivalent to smagic("{op1, op2 ? x}").

Functions

string_ops_alias(): string_ops alias with custom defaults

See also

Author

Laurent R. Berge

Examples

# data on car models
cars = row.names(mtcars)

# let's get the brands starting with an "m"
string_ops(cars, "'i/^m'get, x, unik")
#> Error in string_ops(cars, "'i/^m'get, x, unik"): could not find function "string_ops"

# Explainer:
# 'i/^m'get: keeps only the elements starting with an m,
#            i/ is the 'regex-flag' "ignore" to ignore the case
#            ^m means "starts with an m" in regex language
# x: extracts the first pattern. The default pattern is "[[:alnum:]]+"
#    which means an alpha-numeric word
# unik: applies unique() to the vector
# => see help in ?smagic for more details on the operations

# let's get the 3 largest numbers appearing in the car models
string_ops(cars, "'\\d+'x, rm, unik, num, dsort, 3 first")
#> Error in string_ops(cars, "'\\d+'x, rm, unik, num, dsort, 3 first"): could not find function "string_ops"

# Explainer:
# '\d+'x: extracts the first pattern, the pattern meaning "a succession"
#          of digits in regex language
# rm: removes elements equal to the empty string (default behavior)
# unik: applies unique() to the vector
# num: converts to numeric
# dsort: sorts in decreasing order
# 3 first: keeps only the first three elements