String operations and interpolation: magic edition

The stringmagic package introduces a new language tailored to create complex character strings. Use the function smagic to interpolate a variable inside a character string and to apply, on-the-fly, any arbitrary string operation.

Details

It supports over 50 basic operations, seamless customization, nesting, pluralization, and much more. Writing complex character strings has never been that easy!

To get an overview of the package: take the tour. To get started with smagic, see the vignette.

This package also enhances regular string functions (like grep and co) with specialized functions to:

detect combinations of regular expressions

chain basic string operations

clean character string vectors

See more in the dedicated vignette.